# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Express TypeScript Boilerplate API
  description: |
    Production-ready Express.js TypeScript boilerplate for blockchain microservices.

    This API provides comprehensive user management, authentication, and health monitoring
    capabilities with enterprise-grade security features.

    ## Features
    - JWT and API Key authentication
    - Role-based access control (RBAC)
    - Rate limiting with Redis backend
    - Input validation and sanitization
    - Comprehensive health checks
    - Metrics collection
    - Security headers and CORS

    ## Authentication
    This API supports two authentication methods:
    1. **JWT Bearer Token** - For user sessions
    2. **API Key** - For service-to-service communication

    ## Rate Limiting
    All endpoints are rate limited. Limits vary by endpoint and authentication method.
    Rate limit information is provided in response headers.

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in format "keyId.key"
...
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
            version:
              type: string
              example: v1
...
    User:
      type: object
      properties:
        id:
          type: string
          example: clp1234567890abcdef
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: john_doe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        isActive:
          type: boolean
...
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          example: 900
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: VALIDATION_ERROR
                      message:
                        type: string
                        example: Validation failed
                      details:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: email
                            message:
                              type: string
                              example: Invalid email format
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: UNAUTHORIZED
                      message:
                        type: string
                        example: Authentication required
...
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: CONFLICT
                      message:
                        type: string
                        example: Resource already exists
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: RATE_LIMIT_EXCEEDED
                      message:
                        type: string
                        example: Rate limit exceeded
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../base.yaml#/components/schemas/RefreshTokenRequest
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/LoginResponse
        '400':
          $ref: ../base.yaml#/components/responses/ValidationError
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: |
        Retrieve the profile of the currently authenticated user.
        This endpoint allows users to view their own profile information.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Current user profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
    put:
      tags:
        - Users
      summary: Update current user profile
      description: |
        Update the profile of the currently authenticated user.
        Users can only update their own profile information.
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../base.yaml#/components/schemas/UpdateUserRequest
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '400':
          $ref: ../base.yaml#/components/responses/ValidationError
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '409':
          $ref: ../base.yaml#/components/responses/ConflictError
        '429':
          $ref: ../base.yaml#/components/responses/RateLimitError
...